+++ src/utils/_objects/isEqual.ts
@@ -1,0 +1,3 @@
+import _isEqual from 'lodash.isequal'
+
+export const isEqual = _isEqual
