+++ src/utils/index.ts
@@ -1,6 +1,6 @@
 export { Arrays } from './arrays'
 export { Dates } from './dates'
 export { Numbers } from './numbers'
-export { Objects } from './objects'
+export { Objects } from './_objects'
 export { Queue } from './queue'
 export { UUIDs } from './uuids'
