--- src/utils/objects/isEqual.ts
@@ -1,1 +1,0 @@
-export { default as isEqual } from 'lodash.isequal'
